[5] - OpenWalletService created and controller logic extracted
